ajax and etc

+++ src/Controller/PostController.php
@@ -11,10 +11,12 @@ use App\Form\PostType;
 use App\Message\SmsNotification;
 use App\Repository\PostRepository;
 use App\Repository\UserRepository;
+use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Finder\Exception\AccessDeniedException;
@@ -27,6 +29,7 @@ use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\Messenger\MessageBusInterface;
+use Symfony\Contracts\Cache\CacheInterface;
 
 /**
  *
@@ -62,6 +65,9 @@ class PostController extends AbstractController
             $request->query->getInt('page', 1)/*page number*/,
             5/*limit per page*/
         );
+
+
+        //dd($pagination);
 
         return $this->render('post/index2.html.twig', [
             'pagination' => $pagination,
@@ -131,7 +137,6 @@ class PostController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'post_edit', methods: ['GET', 'POST'])]
-    //   #[IsGranted('ROLE_ADMIN')]
     public function edit(Request $request, Post $post, EntityManagerInterface $entityManager): Response
     {
         $this->denyAccessUnlessGranted('ROLE_USER');            //todo
@@ -195,7 +200,7 @@ class PostController extends AbstractController
 
 
         if ((strcmp($user->getUserIdentifier(), $post->getUserId()->getUserIdentifier())) == 0 ||
-        in_array('ADMIN',$user->getRoles(), true)) {
+            in_array('ROLE_ADMIN', $user->getRoles(), true)) {
             $entityManager->remove($post);
             $entityManager->flush();
 
@@ -216,5 +221,32 @@ class PostController extends AbstractController
 //        }
 
         return $this->redirectToRoute('post_index', [], Response::HTTP_SEE_OTHER);
+    }
+
+    #[Route('/check', name: 'check', methods: ['POST', 'GET'])]
+    public function checkActualInfo(Request $request, CacheInterface $cache) //todo long pooling !!!!
+    {
+        $pageLoadTime = (int)$request->getContent();
+
+        //$ts = (new DateTime)->setTimestamp($request->getContent())->format('H:i:s \O\n Y-m-d');
+
+        $cachePool = $cache->get('time_stamp', function () {
+//            $time = new \DateTime();
+//            $time->getTimestamp();
+            return time();
+        });
+
+        //dd('load ' . $pageLoadTime, $cachePool);
+
+//        $comments_last_write_time = Cache::remember('comments-last-write-time', 86400 * 30, function () {
+//            return Carbon::now();
+//        });
+        //return ['is_modified' => ($comments_last_write_time >= $timestamp)];
+        $response = new Response();
+        $response->setContent(json_encode([
+            'is_modified' => ($pageLoadTime - 547 >= $cachePool),
+        ]));
+        $response->headers->set('Content-Type', 'application/json');
+        return $response;
     }
 }
